Refactor blocking platform calls in the vm
When executing calls that potentially block for a long time, like
for example 'accept' on a tcp server socket its imporant to mark
the 'excutor' as paused during the call.
Without marking the executor as paused the garbage collector can
not run during the call.

However allot of care must be taken that any resources are on the
stack while being 'paused' (and not in local variables), otherwise the
gc could delete them.

Also when returning from a 'Paused' call its important to check if the
gc is running atm (if so we need to 'trap' ourselves) and we need to
check if our executor was aborted while we where paused.