Add implicit conv in func templ type inference
Allow implicit conversions in function template type
inference.

For example:
```
fun add{T}(T a, T b)
  a + b
print(add(13.337, 1))
```
Is now legal as int is implicitly convertable to float
so T is inferred as float.

Note: It only checks if further usages are compatible
with the type it has inferred so far (so it checks if int
is compatible with float).

It does not examine all usages to find out if there is a
type that all would be compatible with. In theory this is
possible and we could consider it if we find a usecase.