Fix use after free in ref_future
Caused when ref_future is destructed while
other threads are still waiting on it (which can
happen when tearing down the vm as threads
waiting on a future are considered as paused.

This fix this ref_future now keep tracks of how
many threads are still waiting on it and in this
destructor signals them and waits for them.