Shebang in nova format (#182)
* novasm: Add shebang line to assembly format

Add a shebang line to instruct operating systems that support
it to execute this file with the 'novrt' runtime executable. This
is just an optional feature and does not hinder systems that do
not support it.

NOTE: To avoid hardcoding the absolute path of novrt we
use '/usr/bin/env'  to lookup  the executable. This is
not fully portable but works across many posix systems.

 More info: https://en.wikipedia.org/wiki/Shebang_(Unix)

* novrt: Drop support for implicit 'prog.nova' path

It forced novus assembly files to always have the
'nova' extension, which is not something we want
to enforce in the long term.