Windows install (#185)
* rt: Add infrastructure for meta commands to novrt

If the first argument starts with two dashes it will be interpreted
as a meta-command. This way we can implement commands like
'--version', '--help' and '--install' etc.

Note: 'novrt' will either execute a meta-command, or a novus
executable. Never both.

* novrt: Add windows install and uninstall meta cmds

Adds registry keys to associate '.nx' files with the novrt executable

* novrt: During windows install register .nx as executable

During installation add '.NX' to 'PATHEXT' env var, to let
windows know that it can execute '.nx' files.

* novrt: Win32 expand the PATHEXT env var

Before we would just place the contents of the machine-wide
PATHEXT into your user PATHEXT if you did not have a PATHEXT yet.

Instead we now expand the machine PAHTEXT with your own.

* novrt: Win32 add to 'Path' env var in install cmd

* novrt: Set friendly type name during win32 install

* build: Add convenience install batch scripts

* build: Update win install batch scripts

They now find the novrt.exe from the dir of the script, instead
of the current working directory,

* build: Statically link the win c++ runtime lib

On windows shipping a c++ redistructable is kinda a pain and
the size of embedding the c++ runtime in our executables is
managable. novrt.exe goes from about 100kb to 500 kb.