frontend: Refactor Diag to use the SourceTable
Decreases the memory footprint of diagnostics, but more
importantly it will be more consistent with how we expose
pointers to  source-code in the future.

Note: We lost the feature that unit-tests where verifying
if the correct source-spans where set in the Diagnostic.
This is unfortunate, however those made the tests hare to
write and maintain in general. If we want to test those i think
it would be better to write specialized tests for that.