Move equality operators to library (#195)
* frontend: Tweak func templ instantiation rules

Don't attempt to instantiate a function template if there's
a non-templated function that could be called. Even if it requires
conversions. The reason this makes it easier to specialize templated
functions with non-templated implementations. Currently it often fails
due to required implicit conversions (like int to long).

* prog: Convert equality ops into intrinsics

Allows for more flexiblity in the library, by defining the
equality operators there instead of being build-in.

* std: Implement equality operators

* prog: Remove bool equality

As implemented using int eq anyway.

* vm: Replace LogicInvInt with CheckIntZero

Implementation is identical but the name now implies more
general use.