frontend: Tune the templated type resolving
Before we would always prefer non-templated versions
of the same type. Even if you are calling a constructor for a
templated version. This is problematic if your overloading types
based on amount of template parameters.

The new logic perfers finding a type which you can construct from
your arguments. But only if you don't explicitly provide type paramters.