frontend: Allow fail intrinsic in pure contexts
Reason for this change is in rare situations you can hint
to the compiler that this path is never taken.

It is debatable if failing the program is a side-effect. For example
div by zero results in the same effect (runtime panicing) and yet
still we consider integer devision 'pure'.

Its a slipperly slope and thus we won't make a library wrapper for it
and has to be called using the intrinsics syntax.